feat: pagina de editar cancelar e cadastrar usuarios de adm e alguns bug fix, rotas e etc

+++ app/services/user_service.py
@@ -1,5 +1,6 @@
 from app import db
 from app.models.user import User
+from app.models.reservation import Reservation
 
 class user_service:
 
@@ -65,8 +66,22 @@ class user_service:
     @staticmethod
     def deleteUser(user_id):
         try:
-            user = db.session.query(User).get(user_id)
+            user = User.query.get(user_id)
             if user:
+                # Verificar se há reservas não finalizadas
+                unfinished_reservations = Reservation.query.filter_by(user_id=user_id).filter(
+                    Reservation.status.notin_(['finalizada', 'cancelada'])).all()
+
+                if unfinished_reservations:
+                    print(f"User {user_id} has unfinished reservations.")
+                    return False
+
+                # Excluir todas as reservas finalizadas (se houver)
+                reservations_to_delete = Reservation.query.filter_by(user_id=user_id).all()
+                for reservation in reservations_to_delete:
+                    db.session.delete(reservation)
+
+                # Excluir o usuário fora do loop de reservas
                 db.session.delete(user)
                 db.session.commit()
                 print(f"User deleted: {user}")
@@ -74,6 +89,7 @@ class user_service:
             else:
                 print(f"User {user_id} not found for deletion.")
                 return False
+
         except Exception as e:
             db.session.rollback()
             print(f"Error deleting user {user_id}: {e}")
@@ -123,3 +139,26 @@ class user_service:
         except Exception as e:
             print(f"Error getting user by ID {user_id}: {e}")
             return None
+
+
+    def updateUserAdm(user_id, phone=None, name=None, email=None):
+        try:
+            user = User.query.get(user_id)
+            if user:
+                if phone:
+                    user.phonenumber = phone
+                if name:
+                    user.name = name
+                if email:
+                    user.email = email
+                db.session.commit()
+                print(f"User updated: {user}")
+                return True
+            else:
+                print(f"User {user_id} not found for update.")
+                return False
+        except Exception as e:
+            db.session.rollback()
+            print(f"Error updating user {user_id}: {e}")
+            return False
+
